pyproject.toml
[project]
name = "hse_dd"
authors = [
    {name = "Maskaev Artem"},
    {name = "Anton Gusev"},
]
license = {file = "LICENSE"}
requires-python = ">=3.11"
version="0.0.1"
dependencies = [
    "tensorboard==2.17.0",
    "kagglehub==0.3.4",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "kaggle>=1.5.16",
    "pydantic>=2.0.0",
    "streamlit>=1.29.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "opencv-python>=4.8.0"
]
description = "DD homework"

[build-system]
requires = [
    "setuptools>=45",
]
build-backend = "setuptools.build_meta"

[project.urls]
"Source" = "https://github.com/ye11owSub/hse-dd"


[tool.poetry.group.dev.dependencies]
notebook = "^7.3.2"
ipykernel = "^6.29.5"

